toolkit/components/url-classifier/tests/mochitest/test_allowlisted_annotations.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test the URI Classifier</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>

<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">

<script class="testbody" type="text/javascript">

var Cc = SpecialPowers.Cc;
var Ci = SpecialPowers.Ci;

const {UrlClassifierTestUtils} = ChromeUtils.import("resource://testing-common/UrlClassifierTestUtils.jsm");

// Add https://allowlisted.example.com to the permissions manager
SpecialPowers.addPermission("trackingprotection",
                            Ci.nsIPermissionManager.ALLOW_ACTION,
                            { url: "https://allowlisted.example.com" });

function clearPermissions() {
  SpecialPowers.removePermission("trackingprotection",
                                 { url: "https://allowlisted.example.com" });
  ok(!SpecialPowers.testPermission("trackingprotection",
                                   Ci.nsIPermissionManager.ALLOW_ACTION,
                                   { url: "https://allowlisted.example.com" }));
}

SpecialPowers.pushPrefEnv(
  {"set": [["urlclassifier.trackingTable", "test-track-simple"],
           ["privacy.trackingprotection.enabled", true],
           ["channelclassifier.allowlist_example", true]]},
  test);

function test() {
  SimpleTest.registerCleanupFunction(UrlClassifierTestUtils.cleanupTestTrackers);
  UrlClassifierTestUtils.addTestTrackers().then(() => {
    document.getElementById("testFrame").src = "allowlistAnnotatedFrame.html";
  });
}

// Expected finish() call is in "allowlistedAnnotatedFrame.html".
SimpleTest.waitForExplicitFinish();

</script>

</pre>
<iframe id="testFrame" width="100%" height="100%" onload=""></iframe>
</body>
</html>
